feat: alert setting

+++ project02-teamB-OUR-consumer/project02-teamB-OUR-consumer/Views/StudyView/StudyDetailView.swift
@@ -8,6 +8,11 @@
 import SwiftUI
 import CoreLocation
 
+enum StudyDetailAlert {
+    case delete
+    case normal
+}
+
 struct StudyDetailView: View {
     
     @Environment(\.dismiss) private var dismiss
@@ -21,8 +26,8 @@ struct StudyDetailView: View {
     @State var isShowingReportSheet: Bool = false
     @Binding var isSavedBookmark: Bool
     @State var showAlert: Bool = false
-    @State private var showDeleteAlert: Bool = false
     @State var alertText: String = ""
+    @State var alertCase:StudyDetailAlert = .normal
     
     var body: some View {
         NavigationStack {
@@ -147,6 +152,7 @@ struct StudyDetailView: View {
                                 //MARK: 1 - 내가 작성한 글
                                 if isMyStudy()  {
                                     NavigationLink {
+                                        //TODO: 수정페이지로 이동 왜 안될까
                                         StudyDetailEditView(viewModel: viewModel, study: study)
                                     } label: {
                                         Text("수정")
@@ -158,9 +164,13 @@ struct StudyDetailView: View {
                                     }
                                     Button {
                                         if viewModel.studyDetail.currentMembers.isEmpty {
-                                            showDeleteAlert = true
+                                            alertCase = .delete
+                                            showAlert = true
                                         } else if viewModel.studyDetail.currentMembers.count >= 1 {
-                                            //TODO: 참석자가 한 명 이상이라면 삭제할 수 없다는 알럿
+                                           print("취소못해")
+                                            alertText = "참석자가 있는 스터디는 삭제할 수 없습니다."
+                                            alertCase = .normal
+                                            showAlert = true
                                         }
                                     } label: {
                                         Text("삭제")
@@ -170,48 +180,47 @@ struct StudyDetailView: View {
                                             .background(Color(red: 215 / 255, green: 215 / 255, blue: 215 / 255))
                                             .cornerRadius(5)
                                     }
-                                } else if viewModel.studyDetail.isJoined {
-                                    //MARK: 2 - 이미 참석한 스터디
-                                    Button(action: {
-                                        
-                                    }, label: {
-                                        Text("이미 참석한 스터디입니다.")
-                                            .bold()
-                                            .frame(width: 290, height: 40)
-                                            .foregroundColor(.white)
-                                            .background(.gray)
-                                            .cornerRadius(5)
-                                    }).disabled(true)
-                                    Button {
-                                        //TODO: isSaved 변수 업데이트
-                                        isSavedBookmark.toggle()
-                                        if isSavedBookmark {
-                                            viewModel.updateBookmark(studyID: viewModel.studyDetail.id)
-                                        } else {
-                                            viewModel.removeBookmark(studyID: viewModel.studyDetail.id)
-                                        }
-                                    } label: {
-                                        Image(systemName: isSavedBookmark ? "bookmark.fill" : "bookmark")
-                                            .font(.system(size: 30))
-                                            .frame(width: 60, height: 40)
-                                            .foregroundColor(Color(red: 251 / 255, green: 55 / 255, blue: 65 / 255))
-                                    }
                                 } else {
-                                    //MARK: 3 - 아무 케이스도 해당x
-                                    Button {
-                                        //TODO: 참석 프로세스-디비저장-알럿
-                                        Task {
-                                            await viewModel.joinStudy()
-                                            alertText = "스터디에 참여하였습니다."
-                                            showAlert = true
+                                    if viewModel.studyDetail.isJoined {
+                                        //MARK: 2 - 이미 참석한 스터디
+                                        Button(action: {
+                                            //TODO: 참석 취소를 할까말까
+                                        }, label: {
+                                            Text("이미 참석한 스터디입니다.")
+                                                .bold()
+                                                .frame(width: 290, height: 40)
+                                                .foregroundColor(.white)
+                                                .background(.gray)
+                                                .cornerRadius(5)
+                                        }).disabled(true)
+                                    } else if viewModel.studyDetail.currentMembers.count == viewModel.studyDetail.totalMemberCount {
+                                        //MARK: 3 - 인원이 다 찼을 때 모집마감
+                                        Button(action: {
+                                        }, label: {
+                                            Text("모집 마감된 스터디입니다.")
+                                                .bold()
+                                                .frame(width: 290, height: 40)
+                                                .foregroundColor(.white)
+                                                .background(.gray)
+                                                .cornerRadius(5)
+                                        }).disabled(true)
+                                    } else {
+                                        //MARK: 4 - 아무 케이스도 해당x
+                                        Button {
+                                            Task {
+                                                await viewModel.joinStudy()
+                                                alertText = "스터디에 참여하였습니다."
+                                                alertCase = .normal
+                                                showAlert = true
+                                            }
+                                        } label: {
+                                            Text("참석")
+                                                .bold()
+                                                .frame(width: 290, height: 40)
+                                                .foregroundColor(.white)
+                                                .background(Color(red: 9 / 255, green: 5 / 255, blue: 128 / 255))
+                                                .cornerRadius(5)
                                         }
-                                    } label: {
-                                        Text("참석")
-                                            .bold()
-                                            .frame(width: 290, height: 40)
-                                            .foregroundColor(.white)
-                                            .background(Color(red: 9 / 255, green: 5 / 255, blue: 128 / 255))
-                                            .cornerRadius(5)
                                     }
                                     Button {
                                         //TODO: isSaved 변수 업데이트
@@ -255,6 +264,7 @@ struct StudyDetailView: View {
                     Button(action: {
                         if isAlreadyReported() {
                             alertText = "이미 신고한 스터디입니다."
+                            alertCase = .normal
                             showAlert = true
                         } else {
                             isShowingReportSheet = true
@@ -278,20 +288,21 @@ struct StudyDetailView: View {
         .sheet(isPresented: $isShowingReportSheet) {
             StudyCommentReportView(viewModel: viewModel, isStudy: true)
         }
-        .alert(isPresented: $showAlert) {
-            Alert(title: Text("알림"),
-                  message: Text(alertText),
-                  dismissButton: .destructive(Text("확인")) {
-            })
-        }
-        .alert(isPresented: $showDeleteAlert) {
-            Alert(title: Text("게시물을 삭제하겠습니까?"),
-                  message: Text("게시물을 삭제합니다"),
-                  primaryButton: .destructive(Text("삭제")) {
-                viewModel.deleteStudy(studyID: viewModel.studyDetail.id)
-                dismiss()
-            }, secondaryButton: .cancel(Text("취소")))
-        }
+        .alert(isPresented: $showAlert, content: {
+            if alertCase == .normal {
+                return Alert(title: Text("알림"),
+                      message: Text(alertText),
+                      dismissButton: .destructive(Text("확인")) {
+                })
+            } else {
+                return Alert(title: Text("게시물을 삭제하겠습니까?"),
+                      message: Text("게시물을 삭제합니다"),
+                      primaryButton: .destructive(Text("삭제")) {
+                    viewModel.deleteStudy(studyID: viewModel.studyDetail.id)
+                    dismiss()
+                }, secondaryButton: .cancel(Text("취소")))
+            }
+        })
         .onAppear(){
             Task {
                 await viewModel.makeStudyDetail(study: study)
